// src/app/core/models/regularOrder.model.ts
export class AdjustChargesModel {

  constructor() {
    this.isManual = false;
    this.overrideCommodityID = 0;
    this.overrideCommodityName = '';
    this.overridePriceMethodID = 0;
    this.overridePriceMethodName = '';
    this.overrideRateTypeID = 0;
    this.overrideRateTypeName = '';
    this.overrideRateValue = 0;
    this.isEdited = false;
    this.overrideShowOnBOL = false;
    this.chargeUnit = 1;
    this.fullAmount = this.amount + this.overrideAmount;
    this.IsChargeUnitModified = false;
  }
  materialName: string;
  materialID: number;
  chargeName: string;
  chargeID: number;
  showOnBOL: boolean;
  orderQuantity: number;
  shippedQuantity: number;
  priceMethod: string;
  priceMethodID: number;
  rateValue: number;
  rateTypeName: string;
  rateTypeID: number;
  amount: any;
  commodityName: string;
  commodityID: number;
  isManual: boolean;
  type: string;
  overrideShowOnBOL: boolean;
  overrideRateValue: number = 0;
  overrideCommodityID: number = 0;
  overrideCommodityName: string = '';
  overrideRateTypeID: number = 0;
  overrideRateTypeName: string = '';
  overridePriceMethodID: number = 0;
  overridePriceMethodName: string = '';
  overrideAmount = 0;
  isEdited: boolean = false;
  chargeUnit: number;
  fullAmount: number = this.amount + this.overrideAmount;
  isAutoAddedExclude: boolean = false;

  IsAutoAdded: boolean;
  IsDefault: boolean;
  isModified: boolean;
  IsChargeUnitModified: boolean;
}